+++ .template-lintrc.js
@@ -154,7 +154,8 @@ module.exports = {
     {
       "moduleId": "app/templates/components/flexberry-dropdown",
       "only": [
-        "no-unnecessary-concat"
+        "no-unnecessary-concat",
+        "eol-last"
       ]
     },
     {
@@ -509,12 +510,6 @@ module.exports = {
       ]
     },
     {
-      "moduleId": "tests/dummy/app/templates/ember-flexberry-dummy-multi-list",
-      "only": [
-        "quotes"
-      ]
-    },
-    {
       "moduleId": "tests/dummy/app/templates/ember-flexberry-dummy-suggestion-edit",
       "only": [
         "eol-last",
@@ -651,9 +646,7 @@ module.exports = {
     {
       "moduleId": "tests/dummy/app/templates/log-service-examples/settings-example",
       "only": [
-        "block-indentation",
         "eol-last",
-        "quotes",
         "no-duplicate-attributes",
         "no-inline-styles",
         "no-bare-strings"
@@ -823,6 +816,18 @@ module.exports = {
       ]
     },
     {
+      "moduleId": "tests/dummy/app/templates/components-acceptance-tests/flexberry-groupedit/delete-with-details",
+      "only": [
+        "eol-last"
+      ]
+    },
+    {
+      "moduleId": "tests/dummy/app/templates/components-acceptance-tests/flexberry-objectlistview/delete-with-details",
+      "only": [
+        "eol-last"
+      ]
+    },
+    {
       "moduleId": "tests/dummy/app/templates/components-acceptance-tests/flexberry-objectlistview/computable-field",
       "only": [
         "eol-last"
@@ -854,6 +859,18 @@ module.exports = {
       ]
     },
     {
+      "moduleId": "tests/dummy/app/templates/components-acceptance-tests/flexberry-objectlistview/folv-user-settings",
+      "only": [
+        "eol-last"
+      ]
+    },
+    {
+      "moduleId": "tests/dummy/app/templates/components-acceptance-tests/flexberry-objectlistview/ember-flexberry-dummy-multi-list",
+      "only": [
+        "eol-last"
+      ]
+    },
+    {
       "moduleId": "tests/dummy/app/templates/components-examples/flexberry-button/settings-example",
       "only": [
         "eol-last"
@@ -863,13 +880,6 @@ module.exports = {
       "moduleId": "tests/dummy/app/templates/components-examples/flexberry-checkbox/settings-example",
       "only": [
         "eol-last"
-      ]
-    },
-    {
-      "moduleId": "tests/dummy/app/templates/components-examples/flexberry-datepicker/settings-example",
-      "only": [
-        "eol-last",
-        "quotes"
       ]
     },
     {