+++ addon/components/flexberry-lookup.js
@@ -813,6 +813,12 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
   dropdownSettings: undefined,
 
   /**
+    Additional setting for dropdown.
+    Places autocomplete empty value at the end of values list.
+  */
+  emptyValueAtEnd: false,
+
+  /**
     @method onShowHide
   */
   onShowHide(fixedOnVisible = false) {
@@ -1252,6 +1258,10 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
           _this.onShowHide();
         });
 
+        run(() => {
+          _this.onShowHide();
+        });
+
         let autocompleteDirection = get(_this, 'autocompleteDirection');
         if (autocompleteDirection == 'auto')
         {
@@ -1377,6 +1387,10 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
         run.next(() => {
           _this.onShowHide();
         });
+
+        run(() => {
+          _this.onShowHide();
+        });
       }
     });
   },
@@ -1439,7 +1453,12 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
               });
 
               if (!_this.get('required')) {
-                results.unshift({ name: _this.get('placeholder'), value: null });
+                if (_this.get('emptyValueAtEnd')) {
+                  results.push({ name: _this.get('placeholder'), value: null });
+                } else {
+                  results.unshift({ name: _this.get('placeholder'), value: null });
+                }
+
                 resultArray['null'] = null;
               }