+++ blueprints/flexberry-core/files/__root__/locales/en/translations.js
@@ -11,11 +11,11 @@ $.extend(true, translations, {
 <%= modelsImportedProperties %>
   },
 
-  'application-name': 'Application caption',
+  'application-name': '<%= applicationCaption %>',
 
   forms: {
     loading: {
-      'spinner-caption': 'Loading stuff, please have a cold beer...'
+      'spinner-caption': 'Loading stuff, please wait for a moment...'
     },
     index: {
       greeting: 'Welcome to ember-flexberry test stand!'
@@ -50,7 +50,7 @@ $.extend(true, translations, {
       },
 
       footer: {
-        'application-name': 'Application caption',
+        'application-name': '<%= applicationCaption %>',
         'application-version': {
           caption: 'Addon version {{version}}',
           title: 'It is version of ember-flexberry addon, which uses in this dummy application ' +
@@ -61,8 +61,8 @@ $.extend(true, translations, {
 
       sitemap: {
         'application-name': {
-          caption: 'Application caption',
-          title: 'Application title'
+          caption: '<%= applicationCaption %>',
+          title: '<%= applicationCaption %>'
         },
         'application-version': {
           caption: 'Addon version {{version}}',
@@ -74,7 +74,7 @@ $.extend(true, translations, {
           caption: 'Home',
           title: ''
         },
-<%= enApplicationMenu %>
+<%= ruApplicationMenu %>
       }
     },